Change map back to a staic object, move cleanup exectutor to share class
