+++ bundles/io/org.openhab.io.rest/src/main/java/org/openhab/io/rest/internal/filter/DuplicateBroadcastProtectionFilter.java
@@ -8,15 +8,14 @@
  */
 package org.openhab.io.rest.internal.filter;
 
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
 import javax.servlet.http.HttpServletRequest;
+
 import org.atmosphere.cpr.AtmosphereResource;
 import org.atmosphere.cpr.BroadcastFilter.BroadcastAction.ACTION;
 import org.atmosphere.cpr.PerRequestBroadcastFilter;
 import org.codehaus.jackson.map.ObjectMapper;
+import org.openhab.io.rest.internal.listeners.ResourceStateChangeListener;
+import org.openhab.io.rest.internal.listeners.ResourceStateChangeListener.CacheEntry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -24,7 +23,6 @@ import org.slf4j.LoggerFactory;
 /**
  * This Filter prevents duplicate broadcasts   
  *  
- * @author Dan Cunningham
  * @author Oliver Mazur
  * @since 1.0
  *
@@ -33,19 +31,6 @@ import org.slf4j.LoggerFactory;
 public class DuplicateBroadcastProtectionFilter implements PerRequestBroadcastFilter {
 
 	private static final Logger logger = LoggerFactory.getLogger(DuplicateBroadcastProtectionFilter.class);
-	private static final long CACHE_TIME = 300 * 1000; // 5 mins
-	
-	ConcurrentMap<String, CacheEntry> cachedMessages = new ConcurrentHashMap<String, CacheEntry>();
-
-	public DuplicateBroadcastProtectionFilter(){
-		//clear the uuid cache at a regular interval
-		Executors.newSingleThreadScheduledExecutor().scheduleWithFixedDelay(new Runnable() {
-			@Override
-			public void run() {
-				cleanCache();
-			}
-		}, CACHE_TIME, CACHE_TIME, TimeUnit.MILLISECONDS);
-	}
 	
 	@Override
 	public BroadcastAction filter(Object arg0, Object message) {
@@ -79,7 +64,7 @@ public class DuplicateBroadcastProtectionFilter implements PerRequestBroadcastFi
 			return false;
 		}
 		try{
-			CacheEntry entry = cachedMessages.put(clientId, new CacheEntry(responseEntity));
+			CacheEntry entry = ResourceStateChangeListener.getCachedEntries().put(clientId, new CacheEntry(responseEntity));
 			//there was an existing cached entry, see if its the same
 			if(entry != null){
 				ObjectMapper mapper = new ObjectMapper();
@@ -97,54 +82,4 @@ public class DuplicateBroadcastProtectionFilter implements PerRequestBroadcastFi
 		} 
         return false;
 	}
-	
-	/**
-	 * Clean up expired entries in our cache.
-	 */
-	public void cleanCache(){
-		/*
-		 * This map object will start to collect dead uuid entries over time, but its 
-		 * difficult to know when these uuid's are really not valid anymore.  this
-		 * checks for dead entries before adding any new ones.
-		 */
-		long invalidCacheTime = System.currentTimeMillis() - CACHE_TIME;
-		for(String uuid : cachedMessages.keySet()){
-			if(cachedMessages.get(uuid).getCacheTime() <= invalidCacheTime )
-				cachedMessages.remove(uuid);
-		}
-	}
-	
-	/**
-	 * A CacheEntry object is stored in our cache map to prevent duplicate messages
-	 * @author Dan Cunningham
-	 *
-	 */
-	public class CacheEntry {
-		long cacheTime;
-		Object data;
-		/**
-		 * Create a new CacheEntry object the data to cache
-		 * @param data
-		 */
-		public CacheEntry(Object data) {
-			super();
-			this.data = data;
-			this.cacheTime = System.currentTimeMillis();
-		}
-		/**
-		 * 
-		 * @return the time the entry was cached
-		 */
-		public long getCacheTime() {
-			return cacheTime;
-		}
-		/**
-		 * 
-		 * @return the cached data
-		 */
-		public Object getData() {
-			return data;
-		}
-	}
-
 }
